requirements.txt
pandas
python-dotenv
google-cloud-storage
google-cloud-bigquery
google-api-python-client
apache-airflow==2.10.5
apache-airflow-providers-fab
apache-airflow-providers-celery
psycopg2-binary
great_expectations==1.5.6
sqlalchemy-bigquery